+++ main/model/carbonsources.h
@@ -9,11 +9,11 @@
 
 struct CarbonSource {
     CarbonSource() {}
-    CarbonSource(int newX,int newY, int newAmount)
+    CarbonSource(int newX,int newY, double newAmount)
         : x(newX), y(newY), amount(newAmount) {}
     int x;
     int y;
-    float amount;
+    double amount;
 };
 
 class CarbonSourceCollection {
@@ -21,15 +21,15 @@ class CarbonSourceCollection {
         CarbonSourceCollection();
         CarbonSourceCollection(int newX, int newY);
 
-        void addSource(int x, int y, float ammount);
+        void addSource(int x, int y, double ammount);
         void addSource(const CarbonSource & source);
         void addSources(const QVector<CarbonSource> & sources);
         void addSources(const CarbonSourceCollection & collection);
 
-        void removeSourcesPercent(float percent);
+        void removeSourcesPercent(double percent);
 
         const QVector<CarbonSource> * getSources() const;
-        const QVector<CarbonSource> getSourcesPercentage(float percent) const;
+        const QVector<CarbonSource> getSourcesPercentage(double percent) const;
 
     private:
         int x;
